Verify the order of RawResourceClient callbacks in DocumentThreadableLoader
The errors in the order of RawResourceClient callbacks potentially leads to
CORS bypassing, especially in DocumentThreadableLoader.
This CL introduces RawResourceClientStateChecker to check the order of
RawResourceClient callbacks, assuming that the RawResourceClient is added
to at most one RawResource, and applies it to DocumentThreadableLoader.

BUG=633696, 640291

Review-Url: https://codereview.chromium.org/2020313002
Cr-Original-Commit-Position: refs/heads/master@{#413990}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 48136fcd4b053686ce2f1f98e64b00a064907ec4