Add a switch to TestRunner to enable/disable mock spell checker
Most layout tests are not related to spell checker, and we do not want
to see failures due to appearance/disappearance of spelling markers in
their pixel results.

This patch adds a switch to TestRunner (more specifically, its
SpellCheckClient) so that the mock spell checker returns non-empty result
only when the switch is on. It is off by default, and hence, spelling
markers do not appear in unrelated tests. Spell checking tests are modified
to explicitly set this switch on so their results remain the same as before.

The following tests are rebaselined to remove the spelling markers from their
pixel results:
- compositing/overflow/do-not-paint-outline-into-composited-scrolling-contents.html
- editing/execCommand/5569741.html
- editing/inserting/5418891.html
- editing/inserting/insert-div-027.html
- editing/inserting/paragraph-separator-01.html
- editing/inserting/paragraph-separator-02.html
- editing/inserting/paragraph-separator-03.html
- editing/inserting/return-key-with-selection-003.html
- editing/pasteboard/cut-text-001.html
- editing/pasteboard/undoable-fragment-removes.html
- editing/selection/move-backwords-by-word-001.html
- editing/selection/move-by-word-001.html
- editing/style/block-styles-007.html
- fast/dom/blur-contenteditable.html
- fast/dom/focus-contenteditable.html
- fast/forms/text/input-double-click-selection-gap-bug.html
- fast/inline-block/14498-positionForCoordinates.html
- fast/repaint/inline-outline-repaint.html
- virtual/prefer_compositing_to_lcd_text/compositing/overflow/do-not-paint-outline-into-composited-scrolling-contents.html

BUG=639213

Review-Url: https://codereview.chromium.org/2270293003
Cr-Original-Commit-Position: refs/heads/master@{#414042}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 2fa3e76400d5f0fcc39eea19f9818599a3ea5016