Sync XHR on Worker should not crash when response arrives immediately
A WorkerThreadableLoader::MainThreadLoaderHolder instance is created on the
main thread and its reference is sent to the worker thread. The current
implementation sends the reference after the loading starts (on the main
thread), but it's problematic because the loading may exit immediately and
hence it breaks WorkerThreadableLoader::WaitableEventWithTasks's assumption.

BUG=637980

Review-Url: https://codereview.chromium.org/2250203003
Cr-Original-Commit-Position: refs/heads/master@{#414302}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 4f2cec77114821d5ff05041da5c6d969f6a16ab8