Stop SpellChecker::chunkAndMarkAllMisspellingsAndBadGrammar from using TextCheckingParagrah
SpellChecker::chunkAndMarkAllMisspellingsAndBadGrammar() does not really need
a TextCheckingParagraph as its parameter; an EphemeralRange of text to check is
sufficient.

Hence, this patch changes its parameter from a TextCheckingParagraph to an
EphemeralRange, simplifying code and improving performance by eliminating the
overhead for finding the paragraph containing the text to be checked.

BUG=n/a
TEST=n/a; no visible change

Review-Url: https://codereview.chromium.org/2273453003
Cr-Original-Commit-Position: refs/heads/master@{#414328}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 41165826c11cc0b7f1457b4c455d7e2eb2de0819