Move createCrossThreadTask to ExecutionContextTask.h
Move createCrossThreadTask to ExecutionContextTask.h so that
both same thread and cross thread versions of createTask are in the
same header file.
Currently createCrossThreadTask is in CrossThreadTask.h,
but the implementation is a few lines of code, so merging it to
ExecutionContextTask.h and removing CrossThreadTask.h.

BUG=638839

Review-Url: https://codereview.chromium.org/2264543002
Cr-Original-Commit-Position: refs/heads/master@{#414390}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 84f5959aa09e9841e5aea9993247b46b42a60d9a