Cleanup and refactor root scroller machinery.
Some cleanup and restructuring in anticipation of a CL to make root scrollers
work across iframes:

-Removed ChildViewportScrollCallback

This was essentially a no-op ViewportScrollCallback that would be used inside
iframe RootScrollerControllers to make everything more consistent. On reflection,
I think this was a bad choice and we'll just have one ViewportScrollCallback that
gets moved between frames. I've removed ChildViewportScrollCallback and folded
RootViewportScrollCallback into ViewportScrollCallback. RootScrollerController's
m_viewportApplyScroll is now NULL for child frames.

- Initialize ViewportScrollCallback in RootScrollerController

This was in FrameView because FrameView previously didn't expose the
RootFrameViewport which is needed in this callback.  FrameView had to expose
that for other reasons so this cleans things up a little and keeps
RootScrollerController more encapsulated.

- Remove isViewportScrollCallback method on Document

RootScrollerController is now exposed so there's no sense in growing Document's
interface footprint.

- Add m_currentViewportApplyScrollHost

This removes the need for effectiveRootScroller to be the element that has the
apply scroll callback and foreshadows changes coming in a CL to split the
RootScrollerController responsibilities across two classes.

BUG=505516

Review-Url: https://codereview.chromium.org/2273163002
Cr-Original-Commit-Position: refs/heads/master@{#414489}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: db5e0d5832ca0259869c57090a5bb6eee0292205