Make W3CTestExpectationsTest not subclass W3CTestExpectations.
Reason: It's unnecessary and uncommon for test case classes to subclass
the class that they're testing. Dre added this when he added this file,
and I didn't notice this in code review at the time.

This is a refactoring CL, no behavior changes.

Review-Url: https://codereview.chromium.org/2276413002
Cr-Original-Commit-Position: refs/heads/master@{#414626}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ad96d727244b9c5d67772638cf431da338ca9ed4