Stop supporting legacy keywords in Referrer-Policy header
https://github.com/w3c/webappsec-referrer-policy/pull/66 removed support
for legacy keywords in the Referrer-Policy header. There is no need to
support these legacy keywords because the Referrer-Policy header is
brand new (not shipped yet) and there is no legacy usage.

In the course of implementing this, I realized that we do in fact
support legacy keywords for 'referrerpolicy' content attributes. This is
not necessary per the spec, but I left it as-is since we should measure
usage before removing support for legacy keywords in 'referrerpolicy'
attributes.

BUG=641128

Review-Url: https://codereview.chromium.org/2278823004
Cr-Original-Commit-Position: refs/heads/master@{#414903}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9517b0e7af83a9c482e1a02dd8c31d146caf9b37