Fix CSS reference filters with negative transformed children.
The software raster path for filters had always used the
bounds top-left corner as the origin for filters. This
doesn't work if the filtered primitive contains a
negative-transformed child, since the saveLayer() bounds
should be negative (non-zero) with respect to the
primitive.

The solution is to pass through the origin explicitly, and
translate to that rather than the bounds origin before
applying the filter. This will determine the origin for
crop rects, lighting positions, etc.

It turns out that there are certain CSS properties (e.g., "outline") which affect the size of the layer, but are not
supposed to affect layout. In cc, we had previously been
using the layer's src_rect origin to offset filter
parameters in order to compensate for clipping, etc, but
this is insufficient to handle "outset", which shows up
in the layer position but shouldn't affect filter origin.

The offset that we want was already being computed
for us in GraphicsLayer::setOffsetFromLayoutObject()),
but not forwarded to cc. I've done so, with the
attendant changes to WebLayer, WebLayerImpl, mojo bindings, struct_traits, cc::Layer, cc::LayerImpl, EffectNode,
RenderSurfaceImpl, and RenderPassDrawQuad. (I wish I
was kidding.)

BUG=638091
R=ajuma@chromium.org, chrishtr@chromium.org, mbarbella@chromium.org, sky@chromium.org, wkorman@chromium.org
TBR=sky@chromium.org
CQ_INCLUDE_TRYBOTS=master.tryserver.blink:linux_precise_blink_rel

Review URL: https://codereview.chromium.org/2278273004 .

Cr-Original-Commit-Position: refs/heads/master@{#415043}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5941ecb0aea741b27fea283065b25428588dbe43