Make document.rootScroller work properly across iframes.
This patch makes rootScroller work across iframes in a composable way. The main
thrust is that the RootScrollerController class is aware only of local issues.
Its m_rootScroller and m_effectiveRootScroller members always point to Elements
in the associated Document. The TopDocumentRootScrollerController is responsible
for page-level issues. Including determination and management of the "global"
rootScroller. The "global" root scroller is the Element whose scrolling actually
affects the top controls and produces overscroll effects.

Now, root scrollers chain across iframes. To calculate the global root scroller,
we find the effective root scroller of the root frame. If that's an iframe, we
then descend into the iframe's Document and use its effective root scroller. If
that's an iframe, we descend again, and so on.

BUG=505516

Review-Url: https://codereview.chromium.org/2281603002
Cr-Original-Commit-Position: refs/heads/master@{#415081}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 242c9fb14efa1822a64fcf43cda2c2edef71e406