Set the coordinates of the synthetic touch event correctly in SyntheticGestureTargetAura.
In ConvertLocationToTarget function, the offset added to the event location is not scaled,
so we should apply the screen scale factor to the event location after it has been transformed to
the target. Then we can find the right target to send the event using the correct coordinates.

BUG=634343

Review-Url: https://codereview.chromium.org/2269483002
Cr-Original-Commit-Position: refs/heads/master@{#415119}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 34bff62e925142091ce98722c7bd447f51b3794b