Support structured clone atop v8::ValueSerializer behind a RuntimeEnabledFeature.
There is further refactoring to be done to clean this up, but this is sufficient
to make things mostly work. Notably this still does a copy into a 16-bit WTF::String
because at present there is a significant amount of other code that assumes this.

BUG=148757

Review-Url: https://codereview.chromium.org/2277403002
Cr-Original-Commit-Position: refs/heads/master@{#415195}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e28e1e567c840430f2262b0df778cc07285a8c64