Set navigationStart correctly for all load types.
Currently for browser-initiated navigations the navigationStart recorded in the
browser process is discarded for "non-standard" load types, e.g. history
navigations and navigations with should_replace_current_entry flag set. As per
the comments in code, this is done to prevent setting navigationStart before
prompting to unload the previous document. Sadly it is quite wrong and only
causes inconsistency in the meaning of navigationStart (what it means depends on
load type, though the spec makes no distinctions between them). And still in
same-process navigations with WebFrameLoadType::Standard we take the
browser-provided navigationStart, then fire beforeunload (probably waiting for
user action on the dialog) and keep the old timestamp as is.

In this CL navigationStart is reset to Now() after calling a beforeunload
handler. On the other hand, it is not reset for load types other than standard,
if beforeunload have been called earlier or the frame being navigated contains
just an untouched initial document (guaranteed to be empty).

Review-Url: https://codereview.chromium.org/2103733004
Cr-Original-Commit-Position: refs/heads/master@{#415431}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: eacad60524fdf286eccc22ebc160385cda2ad99c