Refactor inspector protocol compiles.
This removes the duplication of outputs from core/inspector:protocol_sources to
the compiled version in core:core_generated. A new target is created
core/inspector:protocol that uses get_target_outputs on :protocol_sources to
avoid duplicating the sources list.

To make this easier to implement, the core_config_add and core_config_remove
configs list were moved to core.gni so it can be shared with multiple core
build files.

The inspector_protocol.gni file was changed from listing the sources to
defining a template to run the script. This elimiates some duplicated code
between the two users of this script.

BUG=

Review-Url: https://codereview.chromium.org/2294123002
Cr-Original-Commit-Position: refs/heads/master@{#415777}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: dd219210ff98b771af6a9ed93852dcf8e242a442