Make tests wait until automation is done to prevent leaks
Adding an umbrella test that waits on the promise returned
by inject_input method so that it makes sure the input
automation is done before the whole test passes and
finishes the test. This prevent the leak that was
happening before as the result of closing the test
early.

BUG=619060

Review-Url: https://codereview.chromium.org/2300623004
Cr-Original-Commit-Position: refs/heads/master@{#416025}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 62234589a64896729acb4a53c874e775aefb3daf