IndexedDB: Align exception priorities with Firefox
In cases where methods are called with multiple failure conditions,
there's a choice of what exception to throw. For example, on a cursor
update() call, if the transaction is inactive *and* the transaction is
read-only *and* the cursor is advancing *and* the cursor's store has
been deleted *and* the new value is invalid, one of several types of
exceptions would be thrown. This CL aligns Blink with Gecko in the
handful of cases where they differed so the same type is thrown.

No meaningful web-facing change - users couldn't compatibly expect a
particular exception type in these cases anyway. (Testing also
revealed a bug in Firefox, reported to their IDB team.)

Review-Url: https://codereview.chromium.org/1897253003
Cr-Original-Commit-Position: refs/heads/master@{#416080}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: dce205414aa1bde4125c3129fef24a8dc8ad32cc