Clean up and fix logic around creating ImageBufferSurface and fallbacks
Avoids a crash if a surface is not valid, we always early out in that
case. Repairs the fallback logic to not fallback in the cases where:
- An external surface was provided.
- is3D() is true, aka WebGL.
In the above cases, we should not fallback if the surface we try to use
is not valid.

R=junov@chromium.org
BUG=641896

Review-Url: https://codereview.chromium.org/2285153004
Cr-Original-Commit-Position: refs/heads/master@{#416097}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 3c5a6056fca4642b919fb468c6ca9bfc152d1c76