Rename SecurityOriginCache to URLSecurityOriginMap
SecurityOriginCache is needed to implement "the origin of the Blob URL"
defined in the File API spec. It's not a cache but a map which needs to
persist. Rename to less confusing name.

This CL also adds comments and do some refactoring around the logic for
readability.

R=mkwst@chromium.org,tommi@chromium.org
BUG=639787

Review-Url: https://codereview.chromium.org/2268973002
Cr-Original-Commit-Position: refs/heads/master@{#416199}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d809019bba4fa9fb90265a64d311826ab3e356e2