Add some basic sanity tests to RootScroller for cases with remote frames.
Added one test that deals with a remote iframe and another with a remote main
frame. RootScroller doesn't quite work in OOPIF scenarios yet but it's still
behind a flag so that's ok. These are mostly testing that we don't crash in the
normal case and that scrolling works ok.

BUG=641775

Review-Url: https://codereview.chromium.org/2290173004
Cr-Original-Commit-Position: refs/heads/master@{#416364}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 4e2f14c7363fe96556aff4f1e4ab44906ddbcd67