Call ImageResource::clear() before setStatus(DecodeError)
Previously, we call setStatus(DecodeError) and then clear(), so the invariant
of "an errored Resource shouldn't have encoded/decoded data" is broken for a
short period of time. It is observable in the case of Issue 643200.
This CL fixes this by calling clear() before setStatus(DecodeError).

BUG=643200, 631321

Review-Url: https://codereview.chromium.org/2303213002
Cr-Original-Commit-Position: refs/heads/master@{#416515}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: da27b2240b0658bb3ba8122e87513b04e9adb0ac